## Coursera-DataScience Note to course evaluator: THis is my repo for the Coursera class. I am aware that the name "Coursera-DataScience" does not match the "datasciencecoursera" name as specified in the course project. This is intentional.
I created the repo with a name to match the standards used by my employer before getting to that step in the instructions.
It would have been easy to make a new repo with the specified name, but that seemed pedantic, since the intent of the exercise
was to learn how to create a repo. The naming convention I used is what I am used to, and while I understand the value in
following the course instructions 100%, in this case it didn't seem to be the goal.

I hope you will take this into account when evaluating.